Compile

Description

This module acts as the project's compiler holding all functions necessary to transform any statement, arithmetic expression or boolean expression into the respective valid set of instructions

Documentation

data Inst

Data type that holds all valid instructions

Constructors

Push Integer

Add

Mult

Sub

Tru

Fals

Equ

Le

And

Neg

Fetch String

Store String

Noop

Branch Code Code

Loop Code Code

data Aexp

Data type that holds all possible representations of an arithmetic expression

Constructors

IntLit Integer
Represents an integer literal, e.g. 1

VarLitA String
Represents a variable in an arithmetic expression, e.g. 1 + x

VarlitAVectorFetch Aexp String
Represents fetching a value from a vector

AddExp Aexp Aexp
Represents an addition between two arithmetic expressions, e.g. 1 + 2

SubExp Aexp Aexp
Represents a subtraction between two arithmetic expressions, e.g. 1 - 2

MultExp Aexp Aexp
Represents a product between two arithmetic expressions, e.g. 1 * 2

data Bexp

Data type that holds all possible representations of an boolean expression

Constructors

BoolLit Bool
Represents a boolean literal, e.g. True

VarLitB String
Represents a variable in a boolean expression, e.g. 1 and x

AndExp Bexp Bexp
Represents the and logical operation between two boolean expressions, e.g. True and False

NegExp Bexp
Represents the not logical operation between two boolean expressions, e.g. neg True

EquExpBool Bexp Bexp
Represents the equality between two boolean expression, e.g. True == False

EquExpInt Aexp Aexp
Represents the equality between two arithmetic expression, e.g. 1 + 2 == 3

LeExp Aexp Aexp
Represents the less than operation between two arithmetic expressions, e.g. 2 <= 3

data Stm

Data type that holds all valid program statements

Constructors

StoreStmA String Aexp
Represents an integer assignment, e.g. x := 1 + 2

VarlitAVectorStore Aexp Aexp String
Represents storing a value from a vector

StoreStmB String Bexp
Represents a boolean assignment, e.g. x := True and False

ParenthStm [Stm]
Represents a chain of statements between parentheses, e.g. (x := 2; y := True and False)

IfStm Bexp Stm Stm
Represents an if-then-else statement, e.g. if True then x := 2 else x := 3

WhileStm Bexp Stm
Represents a while statement, e.g. while True do x := x + 1

type Program = [Stm]

Defines a program as a list of statements

type Code = [Inst]

Defines the program's code as a list of instructions

compile :: Program -> Code

Translates a list of statements into the respective valid list of instructions

compileStm :: Stm -> Code

Translates a statement from Stm into the respective valid list of instructions

compA :: Aexp -> Code

Translates an arithmetic expression from Aexp into the respective valid list of instructions

compB :: Bexp -> Code

Translates a boolean expression from Bexp into the respective valid list of instructions