Lexer

Description

This module holds functions necessary to extract the tokens from input strings that hold valid programs. It also rejects any program that contains invalid tokens

Documentation

data Token

Holds all valid tokens that can be used in a program. Each token has the name of the corresponding operator or constant followed by the suffix Tok

Constructors

PlusTok
Represents the + operator for addition.

MinusTok
Represents the - operator for subtraction.

TimesTok
Represents the * operator for multiplication.

OpenTok
Represents the opening parenthesis '(' for grouping expressions.

CloseTok
Represents the closing parenthesis ')' for grouping expressions.

SemiColonTok
Represents the ';' symbol used as end of statement.

CommaTok
Represents the ',' symbol used as a list element separator.

AndTok
Represents the logical and operation.

NotTok
Represents the logical not operation.

BoolEqTok
Represents the boolean equality '=' operator.

IntEqTok
Represents the integer equality == operator.

LeTok
Represents the less than or equal to <= comparator.

AssignTok
Represents the assignment := operator.

WhileTok
Represents the while loop construct.

DoTok
Represents the 'do' statement in loops.

IfTok
Represents the 'if' conditional construct.

ThenTok
Represents the 'then' branch in a conditional.

ElseTok
Represents the 'else' branch in a conditional.

ForTok
Represents the for loop construct.

IntTok Integer
Represents integer literals, e.g., 123.

VarTok String
Represents variable names, e.g., "my_var".

BoolTok Bool
Represents boolean constants, True or False.

OpenSqTok
Represents the opening square bracket '[' for list syntax.

CloseSqTok
Represents the closing square bracket ']' for list syntax.

DollarTok
Represents the $ symbol for list indexing.

ListTok
Represents the list keyword for construction.

AssignPlusTok
Represents the += compound assignment operator.

AssignSubTok
Represents the -= compound assignment operator.

AssignProdTok
Represents the *= compound assignment operator.

lexer :: String -> [Token]

Translates a program in a string to a list of corresponding valid tokens.

If an invalid token is found an error is displayed.

getWordToken :: String -> Token

Transforms a word into its corresponding token. The word may be one of the reserved keywords ("not", "and", etc.) or a valid name for a variable, that is, a word that starts with an undercased letter followed by any number of letters, digits and underscores (_).