Main

Description

This module holds the primary functions used for the whole program, that is, the functions responsible for parsing a string containing the program's input and running a list of instructions built from the parser.

Documentation

type State = Map String (Either Bool Integer)

Defines the storage of the program, where the values of variables are kept. A variable can hold a boolean or integer value

type Stack = Pilha (Either Bool Integer)

Defines the evaluation stack of the program

createEmptyStack :: Stack

Builds an empty evaluation stack to be used throughout the program

createEmptyState :: State

Builds an empty map to be used as the storage throughout the program

showEither :: (Show a, Show b) => Either a b -> String

Prints the left or right value associated with an Either data variable

stack2Str :: Stack -> String

Builds a string that represents the contents of the evaluation stack that is given as input. The leftmost value of the string represents the value at the top of the evaluation stack, while the rightmost value represents the value at the bottom of the stack.

state2Str :: State -> String

Builds a string that represents the contents of the program's storage. Its values are variable-value pairs that are organized in alphabetical order of the variable name.

run :: (Code, Stack, State) -> (Code, Stack, State)

Evaluates the instruction at the head of the list of instructions, which must be a valid instruction from the Inst data type. The instruction updates the evaluation stack and the program's storage accordingly so that it can be used by the next instruction.

parse :: String -> Program

Parses a string that represents a program, tokenizing it and translating it into the correct statements from the Stm, Aexp and Bexp data types

buildData :: [Token] -> Program

Translates a list of token to the correct statements from the Stm, Aexp and Bexp data types

getStatement :: [Token] -> (Stm, [Token])

Builds a valid statement from the Stm data type with the next tokens that appear in the token list given as input

getForStatement :: [Token] -> (Stm, [Token])

Builds a valid WhileStm with the next tokens that appear in the token list given as input

This function handles the casa where the WhileStm is built from a ForTok

getParenthStatement :: [Token] -> (Stm, [Token])

Builds a valid ParenthStm with the next tokens that appear in the token list given as input, verifying that the next token that is parsed in a OpenTok

getStatement' :: [Token] -> ([Stm], [Token])

Builds a list of valid statements with the next tokens that appear in the token list given as input, until a CloseTok is found. This fetches all statements that are between a pair of correctly matched parentheses.

getStoreStatement :: [Token] -> (Stm, [Token])

Builds a valid StoreStmA or StoreStmB with the next tokens that appear in the token list given as input. The parser identifies if the statement assigns a boolean or integer value to the chosen variable.

getIncrPlusStatement :: [Token] -> (Stm, [Token])

Builds a valid StoreStmA with the next tokens that appear in the token list given as input. This effectively stores the sum between the value of the variable of the next token in the list and an arithmetic expression.

getIncrMultStatement :: [Token] -> (Stm, [Token])

Builds a valid StoreStmA with the next tokens that appear in the token list given as input. This effectively stores the subtraction between the value of the variable of the next token in the list and an arithmetic expression.

getIncrMinusStatement :: [Token] -> (Stm, [Token])

Builds a valid StoreStmA with the next tokens that appear in the token list given as input. This effectively stores the product between the value of the variable of the next token in the list and an arithmetic expression.

isBool :: [Token] -> Bool

Verifies if a token that represents a boolean constant or a boolean operator is found before the next SemiColonTok is found in the next tokens that appear in the token list given as input. This function is used by getStoreStatement to verify if a store statement assign a boolean or integer value to a certain variable.

getStoreDefaultList :: [Token] -> (Stm, [Token])

Builds a valid ParenthStm with the next tokens that appear in the token list given as input, containing a sequence of statements to store a list full of zeros in a variable.

buildDefaultList :: String -> Integer -> [Stm]

Builds the needed statements to store a list with a certain length given as input in a variable whose name is also specified in the input.

getStoreListStatement :: [Token] -> (Stm, [Token])

Builds a valid ParenthStm with the next tokens that appear in the token list given as input, containing a sequence of statements to store a list with integer values in a variable.

buildStoreStatements :: [Token] -> String -> Int -> ([Stm], [Token])

Builds the needed statements to store a list with arithmetic expression in a variable whose name is given as input.

getIfStatement :: [Token] -> (Stm, [Token])

Builds a valid IfStm with the next tokens that appear in the token list given as input.

getWhileStatement :: [Token] -> (Stm, [Token])

Builds a valid WhileStm with the next tokens that appear in the token list given as input.

This function handles the casa where the WhileStm is built from a WhileTok.

getUpdateVectorStatement :: [Token] -> (Stm, [Token])

Builds a valid VarlitAVectorStore to update the contents of a list with the next tokens that appear in the token list given as input

getAexp :: [Token] -> (Aexp, [Token])

Builds a valid Aexp with the next tokens that appear in the token list given as input.

This translations takes the precedence of parenthesis and of the arithmetic operators used in the expression into account.

parseSumSub :: [Token] -> Maybe (Aexp, [Token])

Builds a valid AddExp or SubExp with the next tokens that appear in the token list given as input.

parseProd :: [Token] -> Maybe (Aexp, [Token])

Builds a valid MultExp with the next tokens that appear in the token list given as input.

parseIntVarPar :: [Token] -> Maybe (Aexp, [Token])

Builds the correct statement to either fetch a value from a list, or parse a variable or an integer

This functions also handles the use of parenthesis in arithmetic expressions.

getBexp :: [Token] -> (Bexp, [Token])

Builds a valid Bexp with the next tokens that appear in the token list given as input.

This translations takes the precedence of parenthesis and of the boolean operators used in the expression into account.

parseAnd :: [Token] -> Maybe (Bexp, [Token])

Builds a valid AndExp with the next tokens that appear in the token list given as input.

parseBoolEq :: [Token] -> Maybe (Bexp, [Token])

Builds a valid EquExpBool with the next tokens that appear in the token list given as input.

parseNot :: [Token] -> Maybe (Bexp, [Token])

Builds a valid NegExp with the next tokens that appear in the token list given as input.

This function allows the chaining of "not" statements without the use of parenthesis such as "not not True" which evaluates to True.

parseIntEqLe :: [Token] -> Maybe (Bexp, [Token])

Builds a valid EquExpInt or LeExp with the next tokens that appear in the token list given as input.

parseBoolVarPars :: [Token] -> Maybe (Bexp, [Token])

Builds a valid BoolLit or VarLitB with the next tokens that appear in the token list given as input.

This functions also handles the use of parenthesis in boolean expressions.

testAssembler :: Code -> (String, String)

testParser :: String -> (String, String)

main :: IO ()