Map

Description

This module Map provides a simple binary search tree based map (dictionary) data structure. It allows to perform operations like insertion, and lookup in logarithmic time.

Type

data Map k v

A binary search tree based map, where k is the key and v is the value.

Construction

empty :: Map k v

Creates an empty map.

Operations

insert :: Ord k => k -> v -> Map k v -> Map k v

Inserts a key and a value pair into the map. If the key already exists, it updates the value.

Complexity: O(log n) on average.

map2List :: Map k v -> [(k, v)]

Converts the map to a list of key-value pairs using Inorder Traversal

Complexity: O(n)

Query

find :: Ord k => k -> Map k v -> Maybe v

Finds a value in a tree by key. Returns Nothing if the key is not present, otherwise Just value.

Complexity: O(log n) on average.

isEmpty :: Map k v -> Bool

Checks if the map is empty.