Pilha

Description

This module Pilha provides a generic stack (LIFO) data structure. It allows users to perform basic operations.

The module also includes specialized functions for handling stacks containing Either values. get2RightValues and get2LeftValues.

Type

data Pilha a

Type Pilha represents a stack. It is a wrapper around a list.

Construction

empty :: Pilha a

Creates an empty stack.

Operations

push :: a -> Pilha a -> Pilha a

Adds a new element to the top of the stack.

pop :: Pilha a -> Pilha a

Removes the top element from the stack.

Throws: error if the stack is empty.

top :: Pilha a -> a

Returns the top element of the stack.

get2RightValues :: Pilha (Either a b) -> (b, b, Pilha (Either a b))

Retrieves two consecutive Right values from the top of the stack if available.

Returns a tuple of the two values and the remaining stack.

Throws: error if there aren't two Right values at the top of the stack.

get2LeftValues :: Pilha (Either a b) -> (a, a, Pilha (Either a b))

Retrieves two consecutive Left values from the top of the stack if available.

Returns a tuple of the two values and the remaining stack.

Throws: error if there aren't two Left values at the top of the stack.

Query Functions

isEmpty :: Pilha a -> Bool

Checks if a stack is empty.